Add hacky way to avoid regex stuff
Instead of trying to get every case of the word correctly, I instead
searched for every case were the key happend INSIDE of a tag. Then I
replaced all those instances with a dumb phrase like
"HACKYWAYTOSOLVEAPROBLEM" and then did a find/replace on all the rest of
the times the key matched in the page. Then I substituted the
"HACKYWAYTOSOLVEAPROBLEM"s on the page with their original words. Really
dumb way of doing things that I was just experimenting with.